Incorporate lwlrap into evaluation script
